Reorg code and add entry point.
